requirements.txt
streamlit
openai
langchain
pymupdf
python-dotenv
tiktoken
markdown2
xhtml2pdf
streamlit-tree-select